[make_sfx()] Add -r/--reproducible flag to make reproducible archives
This flag requires GNU tar 1.28 or later and uses the guidelines
from <https://reproducible-builds.org/docs/archives/#full-example>
(and additionally uses the oldgnu archive format) to make the embedded
tar archive reproducible.  The `$SOURCE_DATE_EPOCH` environment variable
is supported with a default value of 0.  The loader is still the rundir
script itself and will still change if different versions of rundir are
used to make the archive.