Export $ARGV0 and (AppRun only) $APPIMAGE and $APPDIR
These are for compatibility with AppRun executables that expect to
be run from an AppImage.

$ARGV0 is the user-specified argv[0] when running as an SFX, and the
user-specified path to the runnable directory otherwise.  It is not
converted to an absolute path, symlinks are not resolved, and it is
exported unconditionally.

$APPIMAGE is only exported when the runfile is called AppRun and when
running as an SFX.  It is the absolute path (with symlinks resolved if
readlink(1) exists) to the SFX file.

$APPDIR is only exported when the runfile is called AppRun (SFX or not).
It is the absolute path to the runnable directory (with symlinks
resolved if readlink(1) exists) to the SFX file.